requirements.txt
# Simple requirements.txt - Only what you need

google-generativeai==0.3.1
streamlit==1.28.1
fastapi==0.104.1
uvicorn==0.24.0
tensorflow==2.14.0
torch==2.1.0
pandas==2.1.3
numpy==1.24.3
plotly==5.18.0
python-dotenv==1.0.0
pydantic==2.5.0
